# openapi-v2.yaml
openapi: 3.0.3
info:
  title: Autumn API v2
  version: 2.0.0
paths:
...
  /api/v2/projects/merge/:
    post:
      operationId: projects_merge
      tags:
      - projects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectMergeRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ProjectMergeRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProjectMergeRequest'
        required: true
      security:
      - basicAuth: []
      - cookieAuth: []
      - tokenAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResource'
          description: ''
...
  /api/v2/reports/hierarchy/:
    get:
      operationId: reports_hierarchy_retrieve
      parameters:
      - in: query
        name: active
        schema:
          type: string
        description: Filter by active state (true or false).
      - in: query
        name: context_ids
        schema:
          type: string
        description: Comma-separated context IDs.
      - in: query
        name: end_date
        schema:
          type: string
        description: Inclusive local date in YYYY-MM-DD format.
      - in: query
        name: exclude_project_ids
        schema:
          type: string
        description: Comma-separated project IDs to exclude.
      - in: query
        name: exclude_subproject_ids
        schema:
          type: string
        description: Comma-separated subproject IDs to exclude.
      - in: query
        name: exclude_tag_ids
        schema:
          type: string
        description: Comma-separated tag IDs to exclude.
      - in: query
        name: note_snippet
        schema:
          type: string
        description: Case-insensitive note substring.
      - in: query
        name: project_ids
        schema:
          type: string
        description: Comma-separated project IDs.
      - in: query
        name: start_date
        schema:
          type: string
        description: Inclusive local date in YYYY-MM-DD format.
      - in: query
        name: subproject_ids
        schema:
          type: string
        description: Comma-separated subproject IDs.
      - in: query
        name: tag_ids
        schema:
          type: string
        description: Comma-separated tag IDs.
      tags:
      - reports
      security:
      - basicAuth: []
      - cookieAuth: []
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportHierarchy'
          description: ''
  /api/v2/reports/tallies/:
    get:
      operationId: reports_tallies_retrieve
      parameters:
      - in: query
        name: active
        schema:
          type: string
        description: Filter by active state (true or false).
      - in: query
        name: by
        schema:
          type: string
          enum:
          - context
          - project
          - status
          - subproject
          - tag
        required: true
      - in: query
        name: context_ids
        schema:
          type: string
        description: Comma-separated context IDs.
      - in: query
        name: end_date
        schema:
          type: string
        description: Inclusive local date in YYYY-MM-DD format.
      - in: query
        name: exclude_project_ids
        schema:
          type: string
        description: Comma-separated project IDs to exclude.
      - in: query
        name: exclude_subproject_ids
        schema:
          type: string
        description: Comma-separated subproject IDs to exclude.
      - in: query
        name: exclude_tag_ids
        schema:
          type: string
        description: Comma-separated tag IDs to exclude.
      - in: query
        name: note_snippet
        schema:
          type: string
        description: Case-insensitive note substring.
      - in: query
        name: project_ids
        schema:
          type: string
        description: Comma-separated project IDs.
      - in: query
        name: start_date
        schema:
          type: string
        description: Inclusive local date in YYYY-MM-DD format.
      - in: query
        name: subproject_ids
        schema:
          type: string
        description: Comma-separated subproject IDs.
      - in: query
        name: tag_ids
        schema:
          type: string
        description: Comma-separated tag IDs.
      tags:
      - reports
      security:
      - basicAuth: []
      - cookieAuth: []
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportTallies'
          description: ''
...
  /api/v2/sessions/:
    get:
      operationId: sessions_list
      parameters:
      - in: query
        name: active
        schema:
          type: boolean
      - in: query
        name: context_ids
        schema:
          type: string
          minLength: 1
      - in: query
        name: end_date
        schema:
          type: string
          format: date
      - in: query
        name: exclude_project_ids
        schema:
          type: string
          minLength: 1
      - in: query
        name: exclude_subproject_ids
        schema:
          type: string
          minLength: 1
      - in: query
        name: exclude_tag_ids
        schema:
          type: string
          minLength: 1
      - in: query
        name: include
        schema:
          enum:
          - note
          type: string
          minLength: 1
        description: '* `note` - note'
      - in: query
        name: limit
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
      - in: query
        name: note_snippet
        schema:
          type: string
          minLength: 1
      - in: query
        name: offset
        schema:
          type: integer
          minimum: 0
          default: 0
      - in: query
        name: project_ids
        schema:
          type: string
          minLength: 1
      - in: query
        name: start_date
        schema:
          type: string
          format: date
      - in: query
        name: subproject_ids
        schema:
          type: string
          minLength: 1
      - in: query
        name: tag_ids
        schema:
          type: string
          minLength: 1
      - in: query
        name: uuid
        schema:
          type: string
          format: uuid
      tags:
      - sessions
      security:
      - basicAuth: []
      - cookieAuth: []
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionListResponse'
          description: ''
    post:
      operationId: sessions_create
      tags:
      - sessions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionTrackRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SessionTrackRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SessionTrackRequest'
        required: true
      security:
      - basicAuth: []
      - cookieAuth: []
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResource'
          description: ''
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResource'
          description: ''
...
  /api/v2/subprojects/merge/:
    post:
      operationId: subprojects_merge
      tags:
      - subprojects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubprojectMergeRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SubprojectMergeRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SubprojectMergeRequest'
        required: true
      security:
      - basicAuth: []
      - cookieAuth: []
      - tokenAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubprojectResource'
          description: ''
...
components:
  schemas:
...
    ByEnum:
      enum:
      - project
      - subproject
      - context
      - status
      - tag
      type: string
      description: |-
        * `project` - project
        * `subproject` - subproject
        * `context` - context
        * `status` - status
        * `tag` - tag
...
    KindEnum:
      enum:
      - subproject
      - residual
      type: string
      description: |-
        * `subproject` - subproject
        * `residual` - residual
...
    NamedResource:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required:
      - id
      - name
...
    ProjectMergeRequest:
      type: object
      properties:
        source_ids:
          type: array
          items:
            type: integer
            minimum: 1
          maxItems: 2
          minItems: 2
        new_name:
          type: string
          maxLength: 255
      required:
      - new_name
      - source_ids
    ProjectResource:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          $ref: '#/components/schemas/StatusEnum'
        description:
          type: string
          readOnly: true
        context:
          allOf:
          - $ref: '#/components/schemas/NamedResource'
          nullable: true
        tags:
          type: array
          items:
            $ref: '#/components/schemas/NamedResource'
          readOnly: true
        start_date:
          type: string
          format: date
          readOnly: true
        last_activity:
          type: string
          format: date
          nullable: true
          readOnly: true
        total_minutes:
          type: number
          format: double
          readOnly: true
        session_count:
          type: integer
      required:
      - context
      - description
      - id
      - last_activity
      - name
      - session_count
      - start_date
      - status
      - tags
      - total_minutes
    ProjectSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required:
      - id
      - name
    ReportHierarchy:
      type: object
      properties:
        projects:
          type: array
          items:
            $ref: '#/components/schemas/ReportHierarchyProject'
      required:
      - projects
    ReportHierarchyChild:
      type: object
      properties:
        kind:
          $ref: '#/components/schemas/KindEnum'
        project_id:
          type: integer
        id:
          type: integer
          nullable: true
        name:
          type: string
          nullable: true
        total_minutes:
          type: number
          format: double
      required:
      - id
      - kind
      - name
      - total_minutes
    ReportHierarchyProject:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        total_minutes:
          type: number
          format: double
        children:
          type: array
          items:
            $ref: '#/components/schemas/ReportHierarchyChild'
      required:
      - children
      - id
      - name
      - total_minutes
    ReportTallies:
      type: object
      properties:
        by:
          $ref: '#/components/schemas/ByEnum'
        entries:
          type: array
          items:
            $ref: '#/components/schemas/ReportTallyEntry'
      required:
      - by
      - entries
    ReportTallyEntry:
      type: object
      properties:
        kind:
          $ref: '#/components/schemas/KindEnum'
        project_id:
          type: integer
        id:
          type: integer
          nullable: true
        name:
          type: string
          nullable: true
        total_minutes:
          type: number
          format: double
        session_count:
          type: integer
      required:
      - id
      - name
      - total_minutes
...
    SessionListResponse:
      type: object
      properties:
        count:
          type: integer
        total:
          type: integer
        sessions:
          type: array
          items:
            $ref: '#/components/schemas/SessionResource'
      required:
      - count
      - sessions
      - total
    SessionResource:
      type: object
      properties:
        id:
          type: integer
        uuid:
          type: string
          format: uuid
          nullable: true
        version:
          type: integer
        project:
          $ref: '#/components/schemas/ProjectSummary'
        subproject_allocations:
          type: array
          items:
            $ref: '#/components/schemas/SubprojectAllocation'
          readOnly: true
        start:
          type: string
          format: date-time
          description: ISO-8601 instant; a timestamp without an offset is interpreted
            as UTC.
        end:
          type: string
          format: date-time
          nullable: true
          description: ISO-8601 instant; a timestamp without an offset is interpreted
            as UTC.
        active:
          type: boolean
          readOnly: true
        auto_stop_at:
          type: string
          format: date-time
          nullable: true
          description: ISO-8601 instant; a timestamp without an offset is interpreted
            as UTC.
        duration_minutes:
          type: number
          format: double
          nullable: true
          readOnly: true
        elapsed_minutes:
          type: number
          format: double
          nullable: true
          readOnly: true
        note:
          type: string
          nullable: true
        commitment_history_unaffected:
          type: boolean
      required:
      - active
      - auto_stop_at
      - duration_minutes
      - elapsed_minutes
      - end
      - id
      - project
      - start
      - subproject_allocations
      - uuid
      - version
    SessionTrackRequest:
      type: object
      properties:
        project_id:
          type: integer
          minimum: 1
        start:
          type: string
          format: date-time
          description: ISO-8601 instant; a timestamp without an offset is interpreted
            as UTC.
        end:
          type: string
          format: date-time
          description: ISO-8601 instant; a timestamp without an offset is interpreted
            as UTC.
        subproject_ids:
          type: array
          items:
            type: integer
            minimum: 1
        note:
          type: string
          nullable: true
        uuid:
          type: string
          format: uuid
        subproject_allocations:
          type: array
          items:
            $ref: '#/components/schemas/SubprojectAllocationWrite'
      required:
      - end
      - project_id
      - start
    StatusEnum:
      enum:
      - active
      - paused
      - complete
      - archived
      type: string
      description: |-
        * `active` - active
        * `paused` - paused
        * `complete` - complete
        * `archived` - archived
    SubprojectAllocation:
      type: object
      properties:
        subproject_id:
          type: integer
        name:
          type: string
        allocation_bp:
          type: integer
      required:
      - allocation_bp
      - name
      - subproject_id
    SubprojectAllocationWrite:
      type: object
      properties:
        subproject_id:
          type: integer
          minimum: 1
        allocation_bp:
          type: integer
          maximum: 10000
          minimum: 1
      required:
      - allocation_bp
      - subproject_id
...
    SubprojectMergeRequest:
      type: object
      properties:
        project_id:
          type: integer
          minimum: 1
        source_ids:
          type: array
          items:
            type: integer
            minimum: 1
          maxItems: 2
          minItems: 2
        new_name:
          type: string
          maxLength: 255
      required:
      - new_name
      - project_id
      - source_ids
    SubprojectResource:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          readOnly: true
        project_id:
          type: integer
        last_activity:
          type: string
          format: date
          nullable: true
          readOnly: true
        total_minutes:
          type: number
          format: double
          readOnly: true
        session_count:
          type: integer
      required:
      - description
      - id
      - last_activity
      - name
      - project_id
      - session_count
      - total_minutes
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    tokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Token-based authentication with required prefix "Token"